Fix outstanding things broken on python 2.7
I.e, with these additional changes, all unittests pass on 2.7.13

Most of the edits are minor tweaks to related to obvious differences between
python2 and python3. E.g

- maximum value of int in hypothesis tests
- checking against the str value of something in unittests

There is one real change:  Converter is modified to simplify howt
the unstruct_strat attributes is used.  The prior behaviour
where the property setter is invoked on __init__ did not work on python 2.7.13

To correct this, the behaviour of this attribute is modified a little:

- the value of the unstrut_strat attribute is now simply a value of the
  UnstructureStrategy enum, and all references to it in cattr package have
  been updated to reflect this

- unstructure_attrs is now a declared method that switches on the value of the
  unstruct_strat attribute