Correct the docs build; add cattr.vendor.typing
Introduces cattr.vendor.typing. This imports the names from the appropriate
vendored version of python, so users of cattr that need to use its vendored
typing can use the exact same import for python2 and python3

Fixes the docs build, also update the documentation and doctests for using
the vendored typing package to refer to cattr.vendor.typing.